+++ src/External/Blockcore.Benchmark/Uint256/UInt256Old.cs
@@ -334,7 +334,7 @@ namespace Blockcore.Benchmark.Uint256.Old
         public int CompareTo(object obj)
         {
             return obj is uint256 v ? CompareTo(v) :
-                   obj is null ? CompareTo(null as uint256) : throw new ArgumentException($"Object is not an instance of uint256", nameof(obj));
+                   obj is null ? CompareTo(null) : throw new ArgumentException($"Object is not an instance of uint256", nameof(obj));
         }
 
         public static bool operator ==(uint256 a, uint256 b)
